put everything of rpc under discord_rpc.rs
